+++ internal/sender/dingtalk/push.go
@@ -3,3 +3,6 @@ package dingtalk
 /*
 	发送前队消息体进行处理
 */
+// func pushJob() {
+
+// }
